+++ NEWS.md
@@ -2,6 +2,12 @@
 
 ## Polars R Package (development version)
 
+### New features
+
+-   In the when-then-otherwise expressions, the last `$otherwise()` is now optional,
+    as in Python Polars. If `$otherwise()` is not specified, rows that don't respect
+    the condition set in `$when()` will be filled with `null` (#836).
+
 ## Polars R Package 0.14.1
 
 ### Breaking changes